Camera Control Script Complete
I finished the script that allows for panning by clicking and dragging,
the mouse wheel to zoom in/out (will be replaced by pinching for mobile
version), and double-clicking for centering (a feature mentioned during
meeting but not in design document).

You will need to attach the script (CameraControl.cs) to the camera. I
had to re-write it once already because github overwrote it (since I am
new to using this) and I lost most of my work, so let me know if there
is anything missing or unclear in the script.