Syncing Changes (Intro Screens)
Attempted to reproduce building site error no luck, but I had a chance
to replace the screens in the intro script with the new ones. I'm not
sure if the display screen is the agreed test size/ratio, since the art
team has been told to use Glen's phone size, which is 960 x 540 pixels.