Intel System Update
Most of the functionatliy of the intel system should be there, but was
unable to test most of it since the connection between the building data
and the link system does not seem to be working (e.g., I attempt to
change the input/output data for the buildings and it does not seem to
make a difference when I run the game). Also, there is a function for it
to resolve events with upgrades, but I was unsure where to put it (it
should be put somewhere where it will be called when an upgrade arrives
at its final destination).