Reorganization of Folders and Some Script Adj.
I moved all the materials to one location
(Assets > Materials > v0) so that we could more easily change them.

I also got rid of the subfolders for the textures so it'd be easy to
edit them all at once, too (except for the prefabs that still have
multiple meshes).

Also got rid of the subfolders for the building prefabs since the
materials and textures are in seperate locations (except for the gas
station for now, needs to be reimported).

Also cleaned up camera control (and other scripts) to get rid of zoom.
Also set option to have camera automatically set within the camera
bounds when game starts.