Continued work on fixing linking bug
Still not fixed, but it's definitely the fact that input on the
deactivated chain are not properly set as no longer deactivated after
the undo occurs. This is probably due to the mutual link...it seems that
when a mutual link is formed, that the inputLinkedTo variable isn't
properly set (observed in ExpLink scene, where the refinery's
inputLinkedTo variable is still set to the city index even after
overriding the link with input from the house).