Possibly Fixed Linking Bug
"If a new link is formed, and doing so causes the outputting building to
deactivate, do not activate the new building"

As specified, when attempting to do such a link, it will cause the game
to undo the action.

Needs to be tested to be verified that it is fixed.