Cleaning Up Script (pt 1)
Removed DefaultBuildings script  almost from usage (it's still in the
folder, it's just not used by any scripts in the game, or at least it
shouldn't...for example, PlaceBuildings uses it still, but the
PlaceBuildings script doesn't look like its in use anymore, either).

I still need to move the class BuildingsOnGridData to the database or
something like that.

I'm going to be doing some cleaning up of the database script, or at
least attempt to, with particular attention to linking...maybe not
changing up anything, but attempting to comment so it's easier to see
what's going on. Please keep an eye out in case this creates any more
bugs.

I also did some other tweaks to remove a lot of Debugs.Log which were
bogging some parts down, as well as started to address some of the
warnings the code has been spouting at us all this time.