Update to Codex Functionality
As it was, each player would have had to save the entire codex within
their player data, but I switched it up so the player only saves a list
of strings, which represent each codex entry they've saved. This also
helps if we ever need to update the codex with new entries.