Objective Icon Update
I think everyone preferred the brown, so that's what I used--I still
think it looks bad with the brown on red (the yellow on brown is less
offensive), so I may try some other color combinations later, or resort
to the original blue. But for now it will be a placeholder.