Rotated Terrain Textures
I rotated essential terrains (grass and water) to be straight on with
the camera. The rock, dirt, asphault, and and sand textures don't really
look like they need to be rotated. The city and sidewalk textures may or
may not need to be rotated depending on how we use them.