Notifications/Pointers only display first time
Using playerprefs, notifications and pointers only display the first
time a level loads (you might have to use the 'quit' button since it
doesn't save playerprefs until then, but I might add in a "save" which
wouldn't be hard.

I do know there is one spot that doesn't seem to be working, but I'm not
sure where some notifications are set? If they are set using the
Inspection Display component or Intel System, it should work...and if
it's using the Tutorial Pointers it should work.