Adding in 'back button' functionality for Android
Search for 'ANDROID BACK BUTTON' to access the additions. What I found
is that the 'escape' key is what's mapped to the back button, so that's
what I used. Needs to be verified on Android still.

The following are currently what you can use the back button for:

- When creating a new profile, pressing the ESC button will return back
to the start screen.
- When on the options menu, return to start screen (and closes
confirmation window if open)
- When on mission select page, returns to start menu
- When on a message on mission select, returns to the list of messages
- Opens pause menu when on HUD
- Resumes game when in pause menu