Improved Codex drag
Originally dependent on the length of the hold, now is dependent on the
distance the player moves to register drag vs. click (which is what I
should have done originally, d'oh).

This is what I believe what the issue with the task "Fix clickable area
of exit button in Codex entry screen" since I had no issues with the
clickable area. My guess was it was having issues differentiating
between a drag and a click.